refactor: simplified and rearranged code
