+++ src/App.css
@@ -4,7 +4,7 @@
 
 :root {
   --BODY-BG-COLOR: #ffffff;
-  --NOTE-BG-COLOR: #c1dcd1;
+  --NOTE-BG-COLOR: #e1eadc;
   --NEW-NOTE-BG-COLOR: #f5f3e9;
   --NEW-NOTE-CTA-BG: #49655e;
   --AUTH-CARD-BG-COLOR: #f1eeeb;
@@ -58,6 +58,7 @@ body {
   padding: 1rem;
   margin: auto;
   box-shadow: 0px 2px 4px var(--SHADOW-GREY), 0px 4px 8px var(--SHADOW-GREY);
+  height: 100%;
 }
 
 .new .note-title {
@@ -68,6 +69,9 @@ body {
   border: none;
 }
 
+.new .note-body{
+  height: 68vh
+}
 .new .note-footer {
   align-self: flex-end;
 }
@@ -130,9 +134,9 @@ body {
   background: none;
 }
 
-.empty-page {
+.empty-page, .error-page {
   position: fixed;
-  top: 65%;
+  top: 55%;
   left: 60%;
   transform: translate(-50%, -50%);
   height: 30rem;
@@ -141,6 +145,19 @@ body {
   color: var(--DEFAULT-GREY)
 }
 
+.error-page{
+  background: var(--PURE-WHITE);
+  height: 90vh;
+  width: 100vw;
+  top: 55%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.error-page img{
+  height: 25rem;
+  margin: 1rem auto;
+}
 .note-header {
   display: flex;
   justify-content: space-between;
@@ -159,7 +176,7 @@ body {
   background: var(--ADD-TAG--CTA-BG);
   border: none;
   border-radius: 4px;
-  padding: 4px 6px
+  padding: 4px 6px;
 }
 
 .dropdown-labels:focus,
@@ -217,11 +234,11 @@ body {
 
 .editor-container {
   position: absolute;
-  height: 40vh;
-  top: 50%;
+  height: 88vh;
+  top: 54%;
   left: 50%;
   transform: translate(-50%, -50%);
-  width: 40vw;
+  width: 80vw;
   z-index: 1;
 }
