+++ src/pages/Authentication/Login.css
@@ -6,7 +6,7 @@
     width: fit-content;
     border-radius: 5px;
     border: none;
-    margin: 2rem auto;
+    margin: 2rem 15rem;
     padding: 3rem;
 }
