+++ src/pages/Authentication/SignUp.css
@@ -4,7 +4,7 @@
     max-width: 40%;
     border-radius: 5px;
     border: none;
-    margin: 1rem auto;
+    margin: 1rem 15rem;
     padding: 3rem 3rem 1rem 3rem;
 }
