+++ src/pages/pages.js
@@ -4,5 +4,6 @@ import MockAPI from "./Mockman";
 import NotesList from "./NotesList";
 import Archive from "./Archive";
 import Trash from "./Trash";
+import Page404 from "./Page404";
 
-export {Login, SignUp, MockAPI, NotesList, Archive, Trash}
+export {Login, SignUp, MockAPI, NotesList, Archive, Trash, Page404}
